Move the Kubernetes client into a library file
