+++ lib/msf/core/post_mixin.rb
@@ -24,7 +24,7 @@ module Msf::PostMixin
     )
 
     register_options( [
-      Msf::OptInt.new('SESSION', [ true, "The session to run this module on." ])
+      Msf::OptInt.new('SESSION', [ true, 'The session to run this module on' ])
     ] , Msf::Post)
 
     # Default stance is active
